chore: relicense under MIT and add a README
It's clear why we need a README, it might be not so clear why we should
use MIT;-)

I choose that license based on the assumption that the costs of improvements
to an open data tool occur on the analysis and not the implementation
part. If published, this derivative work by others could be built upon
also without access to the forks sources. And the assumption, that
the democraphics of the user base reduces the need for copy left enforcement.

MIT shall it be \o/